backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# AI
openai==1.12.0
httpx==0.27.0

# ChromaDB (HTTP client only — connects to ChromaDB container)
chromadb-client==0.5.23

# Validation & Security
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bleach==6.1.0

# Logging
structlog==24.1.0

# Rate Limiting
slowapi==0.1.9

# Utils
python-dotenv==1.0.1
tenacity==8.2.3
